Disabled daemon code causing exceptions, separated limit disk space preference, added current usage info in Preferences dialog

+++ Aural/CommandExecutor.swift
@@ -58,7 +58,10 @@ class CommandExecutor {
     
     static func cancel(_ cmd: Command) {
         
-        cmd.process.terminate()
+        if cmd.process.isRunning {
+            cmd.process.terminate()
+        }
+        
         cmd.cancelled = true
         cmd.enableMonitoring = false
     }
